+++ src/gen_bond_labels.cpp
@@ -1,9 +1,9 @@
 #include <vector>
 #include <algorithm> // for std::sort, std::unique
 
-std::vector<double> gen_bond_labels(const std::vector<std::vector<double> >& g0, const std::vector<std::vector<double> >& g1) {
+std::vector<float> gen_bond_labels(const std::vector<std::vector<float> >& g0, const std::vector<std::vector<float> >& g1) {
   
-  std::vector<double> all_labels;
+  std::vector<float> all_labels;
   all_labels.push_back(0.0);
   all_labels.push_back(1.0);
   all_labels.push_back(1.5);
@@ -13,16 +13,16 @@ std::vector<double> gen_bond_labels(const std::vector<std::vector<double> >& g0,
   all_labels.push_back(5.0);
   all_labels.push_back(6.0);
   // Vector to store potential bond labels (extracted from both matrices)
-  std::vector<double> intersection;
+  std::vector<float> intersection;
 
   // Iterate over rows and columns of g0
   for (size_t i = 0; i < g0.size(); ++i) {
     for (size_t j = 0; j < g0[i].size(); ++j) {
-      double current_label = g0[i][j];
+      float current_label = g0[i][j];
 
       // Check if the label exists in each row of g1 (avoid nested loops)
       bool found_in_g1 = false;
-      for (const std::vector<double>& row : g1) {
+      for (const std::vector<float>& row : g1) {
         if (std::find(row.begin(), row.end(), current_label) != row.end()) {
           found_in_g1 = true;
           break;
