+++ src/hood.cpp
@@ -3,7 +3,7 @@ using namespace std;
 // vtx: selected node
 // g: selected graph
 // edge: bond type
-std::vector<int> hood(int vtx, const std::vector<std::vector<double> >& g, double edge) {
+std::vector<int> hood(int vtx, const std::vector<std::vector<float>>& g, float edge) {
     // Return the neighbors of a specified node, with the specified bond type.
     std::vector<int> friends;
     for (std::size_t i = 0; i < g.size(); ++i) {
@@ -11,5 +11,6 @@ std::vector<int> hood(int vtx, const std::vector<std::vector<double> >& g, doubl
             friends.push_back(i);
         }
     }
+
     return friends;
 }
